Move statistics classes
